backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic-settings
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
requests
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator
httpx==0.26.0